uploading first solution draft
I made comments in the code to make clear, what I did. Essentially, I imported all of our data and mushed it into one table and then I did some formating to fit the already existing quality_checked table in the database. I dod not check all tuples of my table, just scrolled through it. It might be good to do some checks, if everything was imported correctly (due to minor mistakes in the quality checked data files from the students). I already appended the table on the database. The output of the existing code underneath does do something (however I am not 100% sure what it gives out, I think it's the total amount of quality checked data points of each hobo that was used in 2020).