Update Ex6.Rmd
I added the join. It worked, but we have now two columns with the hobo_id (one for each table). But we can access the specific columns with night_ind.hobo_id and day_ind.hobo_id

Now, in SQL only the join with the hobo table is missing, I think.